- change the type of num fields in Instruction into int64. - re-generate transportInstruction.pb.go. - add const in encrypt.go. - fix Fragmenter.lastAckSent() uint64 problem. - fix TestFragmenter uint64 problem. - add const in network.go. - rename addSendState() into addSentState(). - add attemptProspectiveResendOptimization() method in TransportSender. - add rationalizeStates(), sendToReceiver() methods in TransportSender. - add sendInFragments(), sendInterval() methods in TransportSender.

+++ network/fragment.go
@@ -237,7 +237,8 @@ func NewFragmenter() *Fragmenter {
 }
 
 func (f *Fragmenter) lastAckSent() uint64 {
-	return f.lastInstruction.AckNum
+	return uint64(f.lastInstruction.AckNum)
+	// return f.lastInstruction.AckNum
 }
 
 // convert Instruction into Fragments slice.
